Create a todo directory under the app directory with many todo files
There were 11 files under /app/user, so we copied most of those over into the new /todo directory. Some of the .spec tests are commented out and will be fixed later. Next we will start making the server side todo files in order for the client side ones to work.