Update filter by status functionality
The html file now allows input for filter by status, where if you write the entire word "complete" or "incomplete", it will filter todos by that. The typescript component file now contains an if statement for handling the status boolean and letting us use strings for input on the front-end. This closes #7, but we still want to continue #4 so that users can click buttons for complete/incomplete instead of having to type the whole word.